+++ tests/test_config_image.py
@@ -343,6 +343,7 @@ def test_reject():
                 'inner_radius' : { 'type' : 'Random', 'min' : 0, 'max': 10 },
             },
             'skip' : '$obj_num == 9',
+            'quick_skip' : '$obj_num == 10',
         },
         'gal' : {
             'type' : 'Convolve',
@@ -390,12 +391,12 @@ def test_reject():
     else:
         logger = galsim.config.LoggerWrapper(None)
 
-    nimages = 10
+    nimages = 11
     im_list = galsim.config.BuildStamps(nimages, config, do_noise=False, logger=logger)[0]
     # For this particular config, only 6 of them are real images.  The others were skipped.
     # The skipped ones are present in the list, but their flux is 0
     fluxes = [im.array.sum(dtype=float) if im is not None else 0 for im in im_list]
-    expected_fluxes = [1289, 0, 1993, 1398, 0, 1795, 0, 0, 458, 0]
+    expected_fluxes = [1289, 0, 1993, 1398, 0, 1795, 0, 0, 458, 0, 0]
     np.testing.assert_almost_equal(fluxes, expected_fluxes, decimal=0)
 
     # Check for a few of the logging outputs that explain why things were rejected.